requirements.txt
antlr4-python3-runtime==4.13.2
mpmath==1.3.0
sympy==1.13.3
